move NewFlightScreen.kt functions to NewFlightViewModel.kt; add set functions to PreferencesInterface.kt
